refactor: remove admin drink write workflow seam
