+++ app/routes/admin.drinks.new.tsx
@@ -1,11 +1,13 @@
 import { data, href } from "react-router";
-import { routeAction } from "#/app/core/route-action.server";
+import { intent, routeAction } from "#/app/core/route-action.server";
 import { getFormErrors } from "#/app/core/utils";
 import { getDb } from "#/app/db/client.server";
 import { purgeDrinkCache } from "#/app/integrations/fastly.server";
 import { deleteImage, uploadImage } from "#/app/integrations/imagekit.server";
+import { drinkDraftSchema } from "#/app/modules/drinks/drinks";
+import { createAdminDrinksWriteService } from "#/app/modules/drinks/drinks.server";
 import { DrinkForm } from "#/app/ui/admin/drink-form";
-import { createAdminDrinkWriteWorkflow } from "#/app/workflows/admin-drink-write.server";
+import { parseCreateDrinkSubmission } from "#/app/workflows/admin-drink-submission.server";
 import type { Route } from "./+types/admin.drinks.new";
 
 export default function NewDrinkPage({ actionData }: Route.ComponentProps) {
@@ -19,24 +21,40 @@ export default function NewDrinkPage({ actionData }: Route.ComponentProps) {
 }
 
 export async function action({ request }: Route.ActionArgs) {
-  const workflow = createAdminDrinkWriteWorkflow({
-    db: getDb(),
-    uploadImage,
-    deleteImage,
-    purgeDrinkCache,
-  });
+  const submission = await parseCreateDrinkSubmission(request);
 
-  const preparation = await workflow.prepareCreate({ request });
-
-  if (preparation.kind === "invalid") {
+  if (submission.kind === "invalid") {
     return data(
       {
-        fieldErrors: preparation.fieldErrors,
-        formErrors: preparation.formErrors,
+        fieldErrors: submission.fieldErrors,
+        formErrors: submission.formErrors,
       },
-      { status: preparation.status },
+      { status: submission.status },
     );
   }
 
-  return routeAction(request, preparation.intent, { formData: preparation.formData });
+  if (!submission.imageUpload) {
+    throw new Error("Create drink submission parser returned ready without an image upload");
+  }
+
+  const imageUpload = submission.imageUpload;
+  const adminDrinksWriteService = createAdminDrinksWriteService({
+    db: getDb(),
+    writeEffects: { uploadImage, deleteImage, purgeDrinkCache },
+  });
+
+  return routeAction(
+    request,
+    intent({
+      schema: drinkDraftSchema,
+      operation: async (draft) =>
+        adminDrinksWriteService.create({
+          draft,
+          imageBuffer: imageUpload.buffer,
+        }),
+      redirectTo: href("/admin/drinks"),
+      toast: { successMessage: "Drink created!" },
+    }),
+    { formData: submission.formData },
+  );
 }
